Fixed Customer repository. Added enumeration for Role.

+++ Exebite.DataAccess/Repositories/CustomerRepository/ICustomerCommandRepository.cs
@@ -3,7 +3,7 @@ using Exebite.Common;
 
 namespace Exebite.DataAccess.Repositories
 {
-    public interface ICustomerCommandRepository : IDatabaseCommandRepository<int, CustomerInsertModel, CustomerUpdateModel>
+    public interface ICustomerCommandRepository : IDatabaseCommandRepository<long, CustomerInsertModel, CustomerUpdateModel>
     {
         Either<Error, bool> UpdateByGoogleId(CustomerUpdateModel customer);
     }
